pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.8.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "langchain-groq>=0.0.1",
    "mcp-use>=0.1.0",
    "httpx>=0.27.0"
]
